split frame comparison component into modules
